+++ package.json
@@ -26,6 +26,8 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
+    "passport": "^0.6.0",
+    "passport-github2": "^0.1.12",
     "pg": "^8.11.3",
     "postcss": "^8.4.29",
     "react": "^18.2.0",
