+++ src/App.jsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import HomePage from "./components/home/HomePage.jsx/";
-import Dashboard from "./components/dashboard/Dashboard.jsx";
-import LoginV2 from "./components/auth/LoginV2.jsx";
-import SignupV2 from "./components/auth/SignupV2.jsx";
+import Dashboard from "./components/dashboard/DashContainer.jsx";
+import Login from "./components/auth/Login.jsx";
+import Signup from "./components/auth/Signup.jsx";
 import InputInstance from "./components/dashboard/InputInstance.jsx";
 
 function App() {
@@ -13,8 +13,8 @@ function App() {
         <Routes>
           <Route path="/home" element={<HomePage />}></Route>
           <Route path="/dashboard" element={<Dashboard />}></Route>
-          <Route path="/" element={<LoginV2 />}></Route>
-          <Route path="/signup" element={<SignupV2 />}></Route>
+          <Route path="/" element={<Login />}></Route>
+          <Route path="/signup" element={<Signup />}></Route>
         </Routes>
       </BrowserRouter>
     </div>
