--- src/components/dashboard/DashRow.jsx
@@ -1,3 +1,0 @@
-export default function QueryRow() {
-  
-}
