--- src/components/dashboard/MOCK_DATA.json
@@ -1,41 +1,0 @@
-[{"time":"13:40:13","name":"Fliptune","type":"M","depth":5,"query":"C1-F1-CE-32-A5-93"},
-{"time":"14:12:23","name":"Feedfire","type":"F","depth":20,"query":"B9-93-BB-FA-C3-F1"},
-{"time":"18:38:14","name":"Blogspan","type":"M","depth":18,"query":"A7-45-D6-81-58-60"},
-{"time":"23:14:14","name":"Linktype","type":"M","depth":11,"query":"D1-AF-F4-9B-5B-D0"},
-{"time":"14:47:25","name":"Mynte","type":"F","depth":3,"query":"7F-7D-7F-8E-44-68"},
-{"time":"3:16:35","name":"Tavu","type":"M","depth":7,"query":"54-1F-8F-1E-E9-8F"},
-{"time":"21:59:48","name":"Avamba","type":"F","depth":3,"query":"06-E6-F8-DA-D6-B9"},
-{"time":"18:41:23","name":"Aimbu","type":"M","depth":14,"query":"99-FF-98-29-A2-DF"},
-{"time":"18:43:17","name":"Blogspan","type":"F","depth":16,"query":"97-EB-38-8A-10-C7"},
-{"time":"22:52:27","name":"Viva","type":"F","depth":13,"query":"CE-66-58-3D-B7-06"},
-{"time":"15:20:54","name":"Thoughtbridge","type":"M","depth":3,"query":"AF-CB-80-2D-FF-1E"},
-{"time":"1:44:02","name":"Photospace","type":"M","depth":14,"query":"EF-4E-A6-8E-31-1A"},
-{"time":"13:58:03","name":"Eidel","type":"M","depth":10,"query":"81-FD-66-03-F4-08"},
-{"time":"10:43:55","name":"Skipstorm","type":"M","depth":18,"query":"24-37-21-C4-18-6C"},
-{"time":"15:04:32","name":"Katz","type":"F","depth":11,"query":"54-67-35-1A-41-7D"},
-{"time":"3:37:23","name":"Youbridge","type":"M","depth":17,"query":"EC-65-D3-D6-70-8A"},
-{"time":"12:55:12","name":"Dynava","type":"F","depth":1,"query":"2C-41-16-A5-4A-61"},
-{"time":"12:39:06","name":"Teklist","type":"M","depth":19,"query":"4A-2B-42-76-BE-16"},
-{"time":"18:58:27","name":"Roombo","type":"F","depth":8,"query":"49-21-11-9A-FE-AC"},
-{"time":"10:40:49","name":"Flashdog","type":"M","depth":9,"query":"BC-E8-96-AA-51-F9"},
-{"time":"18:58:27","name":"Roombo","type":"F","depth":8,"query":"49-21-11-9A-FE-AC"},
-{"time":"10:40:49","name":"Flashdog","type":"M","depth":9,"query":"BC-E8-96-AA-51-F9"},{"time":"14:12:23","name":"Feedfire","type":"F","depth":20,"query":"B9-93-BB-FA-C3-F1"},
-{"time":"18:38:14","name":"Blogspan","type":"M","depth":18,"query":"A7-45-D6-81-58-60"},
-{"time":"23:14:14","name":"Linktype","type":"M","depth":11,"query":"D1-AF-F4-9B-5B-D0"},
-{"time":"14:47:25","name":"Mynte","type":"F","depth":3,"query":"7F-7D-7F-8E-44-68"},
-{"time":"3:16:35","name":"Tavu","type":"M","depth":7,"query":"54-1F-8F-1E-E9-8F"},
-{"time":"21:59:48","name":"Avamba","type":"F","depth":3,"query":"06-E6-F8-DA-D6-B9"},
-{"time":"18:41:23","name":"Aimbu","type":"M","depth":14,"query":"99-FF-98-29-A2-DF"},
-{"time":"18:43:17","name":"Blogspan","type":"F","depth":16,"query":"97-EB-38-8A-10-C7"},
-{"time":"22:52:27","name":"Viva","type":"F","depth":13,"query":"CE-66-58-3D-B7-06"},
-{"time":"15:20:54","name":"Thoughtbridge","type":"M","depth":3,"query":"AF-CB-80-2D-FF-1E"},
-{"time":"1:44:02","name":"Photospace","type":"M","depth":14,"query":"EF-4E-A6-8E-31-1A"},
-{"time":"13:58:03","name":"Eidel","type":"M","depth":10,"query":"81-FD-66-03-F4-08"},
-{"time":"10:43:55","name":"Skipstorm","type":"M","depth":18,"query":"24-37-21-C4-18-6C"},
-{"time":"15:04:32","name":"Katz","type":"F","depth":11,"query":"54-67-35-1A-41-7D"},
-{"time":"3:37:23","name":"Youbridge","type":"M","depth":17,"query":"EC-65-D3-D6-70-8A"},
-{"time":"12:55:12","name":"Dynava","type":"F","depth":1,"query":"2C-41-16-A5-4A-61"},
-{"time":"12:39:06","name":"Teklist","type":"M","depth":19,"query":"4A-2B-42-76-BE-16"},
-{"time":"18:58:27","name":"Roombo","type":"F","depth":8,"query":"49-21-11-9A-FE-AC"},
-{"time":"10:40:49","name":"Flashdog","type":"M","depth":9,"query":"BC-E8-96-AA-51-F9"},
-{"time":"18:58:27","name":"Roombo","type":"F","depth":8,"query":"49-21-11-9A-FE-AC"}]
