--- src/components/home/TableRow.jsx
@@ -1,4 +1,0 @@
-export default function Dashboard() {
-
-  
-}
