+++ src/components/instances/InstanceTable.jsx
@@ -18,15 +18,6 @@ export default function InstaceTable() {
     getData();
   }, []);
 
-  // const tableData = [
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  //   { name: "Secret key", apiKey: "sk-...yZDhr", created: "Aug 19, 2023" },
-  // ];
-
   return (
     <>
       {data.length === 0 ? (
