+++ src/components/navigation/Sidenav.jsx
@@ -1,4 +1,3 @@
-import QuerySearchBar from "../common/QuerySearch";
 import SuspiciousLog from "../dashboard/SuspiciousLog";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEye } from "@fortawesome/free-solid-svg-icons";
@@ -6,18 +5,23 @@ import { faChartLine } from "@fortawesome/free-solid-svg-icons";
 import { faFileLines } from "@fortawesome/free-regular-svg-icons";
 import { useNavigate } from "react-router-dom";
 
+
 import { useState } from "react";
-import data from "./MOCKdb.json";
 
 export default function Sidenav(props) {
-  // const [supiciousLogs, setSuspiciousLogs] = useState([])
 
-  // const renderLogs = suspiciousLogs.map(suspiciousLogs) => {
-  //   return
-  // }
-  // fetch all of the queries
-  // process into suspiciouLogsArr <- have backend take care of this
-  //map over the array of suspicious objects
+
+/** WORK IN PROGRESS - suspicious queries component
+ * remember to import at the top --> import data from './MOCKdb.json'
+ *  const [supiciousLogs, setSuspiciousLogs] = useState([])
+
+  const renderLogs = suspiciousLogs.map(suspiciousLogs) => {
+    return
+  }
+  fetch all of the queries
+  process into suspiciouLogsArr <- have backend take care of this
+  map over the array of suspicious objects
+
   const queries = data.map((query) => {
     return (
       <SuspiciousLog
@@ -28,17 +32,14 @@ export default function Sidenav(props) {
     );
   });
 
+ */
+
+
   return (
     <nav className="flex flex-col px-10 w-20rem h-screen">
-      {/* <div className="border-b mt-10 border-slate-400 py-3">
-        <button>
-          <QuerySearchBar />
-        </button>
-      </div> */}
       <div className="border-slate-400 mt-10 py-3 flex flex-col items-start">
         <h1 className="text-md w-full mb-5 text-slate border-b pb-3">Data</h1>
-        {/* <button>Test</button> */}
-        <details className="mb-5 px-4 py-2 hover:bg-gray-100 hover:font-semibold rounded-md">
+        {/* <details className="mb-5 px-4 py-2 hover:bg-gray-100 hover:font-semibold rounded-md">
           <summary className="m-1 btn text-slate-700 hover:text-primary list-none cursor-pointer">
             <FontAwesomeIcon className="mr-2 " icon={faEye} />
             Suspicious Queries
@@ -46,7 +47,7 @@ export default function Sidenav(props) {
           <ul className="absolute p-2 bg-base-100 w-52 mt-4 space-y-8">
             {queries}
           </ul>
-        </details>
+        </details> */}
         <h1 onClick={() => props.toggleInsight()} className="font-lg mb-2 text-slate-900 px-4 py-4 hover:bg-gray-100 hover:text-primary rounded-md w-full">
           <FontAwesomeIcon icon={faChartLine} /> Insights
         </h1>
