+++ src/components/navigation/[WIP]mockData.js
@@ -1,3 +1,8 @@
+/**
+ * File name: [WIP]mockData
+ * Purpose: open to contribution. starting point for identifying harmful outbound queries. currently a JS file, test as a JSON, and should become obsolete once working
+ */
+
 [{"id": 1, "depth": 3, "latency": "24ms" },
  {"id": 2, "depth": 9, "latency": "108ms" },
   {"id": 3, "depth": 8, "latency": "92ms" },
@@ -8,4 +13,3 @@
   {"id": 8, "depth": 4, "latency": "35ms" },
   {"id": 9, "depth": 2, "latency": "18ms" }]
 
-
